tripleo ci: added release parameter in registering system

+++ tripleo/deploy_all.sh
@@ -32,15 +32,17 @@ function unregister_rhel_system() {
   if [[ "$ENVIRONMENT_OS" == 'rhel' ]] ; then
     cat <<EOF | ssh -T $ssh_opts $ssh_addr
 subscription-manager unregister || true
-su - stack
-if [[ -f stackrc ]] ; then
-  . stackrc
-  while openstack stack list | grep -q overcloud ; do
-    if openstack stack list | grep overcloud | grep -i delete | grep -iq progress ; then
-      openstack stack delete --yes overcloud
-    fi
-    sleep 5
-  done
+if grep -q stack /etc/passwd ; then
+  su - stack
+  if [[ -f stackrc ]] ; then
+    . stackrc
+    while openstack stack list | grep -q overcloud ; do
+      if openstack stack list | grep overcloud | grep -i delete | grep -iq progress ; then
+        openstack stack delete --yes overcloud
+      fi
+      sleep 5
+    done
+  fi
 fi
 EOF
   fi
